+++ Client/src/components/main/ProfileScreen.tsx
@@ -1,56 +1,19 @@
 import useUserInfoQuery from '../../hooks/useUserInfoQuery';
 import Backdrop from '../common/Backdrop';
 import Button from '../common/Button';
-import { useEffect, useRef, useState } from 'react';
+import { useState } from 'react';
 import LoadingBar from '../common/LoadingBar';
 import { Link, useNavigate } from 'react-router-dom';
 import ChangePwTab from './ChangePwTab';
 import ProfileHeader from './ProfileHeader';
-import useUserFeedListQuery from '../../hooks/useUserFeedListQuery';
-import useInfinteScroll from '../../hooks/useInfiniteScroll';
-
-interface UserFeed {
-  feedId: string;
-  body: string;
-  commentCount: number;
-  createdAt: string;
-  statId: number;
-}
+import UserFeedsTab from './UserFeedsTab';
 
 const ProfileScreen = () => {
   const [tab, setTab] = useState<'password' | 'post'>('post');
 
   const { isLoading, data: userInfo } = useUserInfoQuery();
-  const {
-    data: userFeedList,
-    isLoading: isFeedLoading,
-    isFetching,
-    fetchNextPage,
-    hasNextPage,
-  } = useUserFeedListQuery();
 
-  const fetchTriggerRef = useRef<HTMLDivElement>(null);
-  const feedListContainerRef = useRef<HTMLDivElement>(null);
-
-  const userFeedItems = userFeedList?.pages.flatMap((page) => page.data) || [];
-
-  useInfinteScroll({
-    targetEl: fetchTriggerRef.current,
-    hasMore: hasNextPage || false,
-    onIntersect: fetchNextPage,
-  });
-
-  useEffect(() => {
-    feedListContainerRef.current?.scrollTo(0, 0);
-  }, []);
   const navigate = useNavigate();
-
-  if (isFeedLoading)
-    return (
-      <Backdrop>
-        <LoadingBar />
-      </Backdrop>
-    );
 
   if (isLoading)
     return (
@@ -98,38 +61,7 @@ const ProfileScreen = () => {
             </div>
             {/* 탭 내용 */}
             <div className="w-full h-[260px] p-[8px] bg-[#bf916b] rounded-b-[6px]">
-              {/* 마이포스트 */}
-              {tab === 'post' && (
-                <div
-                  className="w-full h-full overflow-y-scroll"
-                  ref={feedListContainerRef}
-                >
-                  {userFeedItems &&
-                    userFeedItems.map((feed: UserFeed) => (
-                      <Link
-                        to={`/feed/${feed.statId}/detail/${feed.feedId}`}
-                        key={feed.feedId}
-                      >
-                        <div className="cursor-pointer font-[Pretendard]  flex flex-col justify-between  text-[10px] w-[90%] h-[3.25rem]  rounded-lg p-2 mx-auto mb-4 border border-black  bg-[#ffc98f]">
-                          <p className="text-[12px] font-bold">
-                            {feed.body.length > 25
-                              ? feed.body.slice(0, 25) + '...'
-                              : feed.body}
-                          </p>
-                          <div className="flex items-center justify-between">
-                            <span>{feed.commentCount} comments</span>
-                            <span>{feed.createdAt.slice(0, 10)}</span>
-                          </div>
-                        </div>
-                      </Link>
-                    ))}
-                  <div
-                    ref={fetchTriggerRef}
-                    onClick={() => fetchNextPage()}
-                  ></div>
-                  {isFetching && <p>Loading...</p>}
-                </div>
-              )}
+              {tab === 'post' && <UserFeedsTab />}
               {tab === 'password' && <ChangePwTab />}
             </div>
           </div>
